:sparkles: 기능 추가 및 멤버 연동
장소 리뷰 추가 기능, 장소 리뷰로 인한 장소 평점 최신화 기능, 대부분의 기능 멤버 연동, 아무튼 기능 추가

+++ BE/src/main/java/com/seoro/seoro/domain/dto/Place/PlaceDto.java
@@ -22,5 +22,5 @@ public class PlaceDto {
 	private String placeLongitude;
 	private List<PlacePhoto> placePhoto;
 	private List<PlaceReviewDto> placeReview;
-	private Long score;
+	private Float score;
 }
